refactor(web/performance): clean ordering
